moved test gc out of main conftest and into subdirectory
